Main menu drop down menu moved to saparete component
